+++ Sources/CohesionKit/Observer/Subscription.swift
@@ -1,0 +1,21 @@
+import Foundation
+
+public class Subscription {
+    public let unsubscribe: () -> Void
+
+    init(unsubscribe: @escaping () -> Void) {
+        var unsubscribed = false
+
+        self.unsubscribe = {
+            if !unsubscribed {
+                unsubscribe()
+            }
+
+            unsubscribed = true
+        }
+    }
+
+    deinit {
+        unsubscribe()
+    }
+}
